+++ src/cgame/cg_info.c
@@ -2053,7 +2053,7 @@ void CG_DemoHelpDraw(void)
  * @param[in] len
  * @return
  */
-char *CG_getBindKeyName(const char *cmd, char *buf, int len)
+char *CG_getBindKeyName(const char *cmd, char *buf, size_t len)
 {
 	int j;
