+++ src/renderer/tr_local.h
@@ -1334,7 +1334,7 @@ qhandle_t RE_RegisterModel(const char *name);
 qhandle_t RE_RegisterSkin(const char *name);
 void RE_Shutdown(qboolean destroyWindow);
 
-qboolean R_GetEntityToken(char *buffer, int size);
+qboolean R_GetEntityToken(char *buffer, size_t size);
 
 float R_ProcessLightmap(byte *pic, int in_padding, int width, int height, byte *pic_out);
