+++ src/renderer/tr_shader.c
@@ -1873,7 +1873,7 @@ static qboolean ParseShader(char **text)
 		}
 		else if (!Q_stricmp(token, "sunshader"))
 		{
-			int tokenLen;
+			size_t tokenLen;
 
 			token = COM_ParseExt(text, qfalse);
 			if (!token[0])
