+++ src/renderercommon/tr_public.h
@@ -119,7 +119,7 @@ typedef struct
 
 	void (*DrawDebugText)(const vec3_t org, float r, float g, float b, const char *text, qboolean neverOcclude);
 
-	qboolean (*GetEntityToken)(char *buffer, int size);
+	qboolean (*GetEntityToken)(char *buffer, size_t size);
 
 	void (*AddPolyBufferToScene)(polyBuffer_t *pPolyBuffer);
 
@@ -178,11 +178,11 @@ typedef struct
 	// won't be freed
 	void (*Hunk_Clear)(void);
 #ifdef HUNK_DEBUG
-	void * (*Hunk_AllocDebug)(int size, ha_pref pref, char *label, char *file, int line);
+	void * (*Hunk_AllocDebug)(size_t size, ha_pref pref, char *label, char *file, int line);
 #else
-	void * (*Hunk_Alloc)(int size, ha_pref pref);
+	void * (*Hunk_Alloc)(size_t size, ha_pref pref);
 #endif
-	void * (*Hunk_AllocateTempMemory)(int size);
+	void * (*Hunk_AllocateTempMemory)(size_t size);
 	void (*Hunk_FreeTempMemory)(void *block);
 
 	// dynamic memory allocator for things that need to be freed
